+++ lib/active_record_streams/message.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require 'securerandom'
+
 module ActiveRecordStreams
   class Message
     def initialize(table_name, action_type, old_image, new_image)
@@ -9,8 +11,20 @@ module ActiveRecordStreams
       @new_image = new_image
     end
 
+    def self.from_json(json)
+      parsed_json = JSON.parse(json)
+
+      new(
+        parsed_json['TableName'],
+        parsed_json['ActionType'],
+        parsed_json['OldImage'],
+        parsed_json['NewImage']
+      )
+    end
+
     def json
       {
+        EventID: SecureRandom.uuid,
         TableName: @table_name,
         ActionType: @action_type,
         OldImage: @old_image,