+++ lib/active_record_streams/publishers/kinesis_stream.rb
@@ -13,17 +13,20 @@ module ActiveRecordStreams
       # @param [String] table_name
       # @param [Enumerable<String>] ignored_tables
       # @param [Hash] overrides
+      # @param [Proc] error_handler
 
       def initialize(
         stream_name:,
         table_name: ANY_TABLE,
         ignored_tables: [],
-        overrides: {}
+        overrides: {},
+        error_handler: nil
       )
         @stream_name = stream_name
         @table_name = table_name
         @ignored_tables = ignored_tables
         @overrides = overrides
+        @error_handler = error_handler
       end
 
       ##
@@ -34,12 +37,12 @@ module ActiveRecordStreams
         return unless (any_table? && allowed_table?(table_name)) ||
                       table_name == @table_name
 
-        client.publish(
-          @stream_name,
-          partition_key(table_name),
-          message.json,
-          @overrides
-        )
+        client.publish(@stream_name, partition_key(table_name),
+                       message.json, @overrides)
+      rescue StandardError => e
+        raise e unless @error_handler.is_a?(Proc)
+
+        @error_handler.call(self, table_name, message, e)
       end
 
       private